refactor: LinkStatus를 shared로 이동하여 의존성 방향 수정 - 도메인이 shared를 참조하도록 변경, 순환 의존성 해결, 아키텍처 일관성 확보
